tc-nios2.c: removed assert in md_number_to_chars
md_number_to_chars gives an assert when running gas if the asm
contains a .8byte or .16byte directive

the nios2 version has an assert in the function.
I checked the arm, ppc and mips versions. They have exactly the same code but
without the assert. Therefore I considered it safe to remove the assert too.

Signed-off-by: Frans Meulenbroeks <fransmeulenbroeks@gmail.com>
Signed-off-by: Tobias Klauser <klto@zhaw.ch>